Enable more compiler warnings

+++ src/common/utils.h
@@ -32,9 +32,9 @@ extern "C" {
 #endif
 
 bool is_power_of_two(uint64_t n);
-int log2_pow2(uint32_t n);
-uint32_t reverse_bits(uint32_t n);
-uint32_t reverse_bits_limited(uint32_t n, uint32_t value);
+uint64_t log2_pow2(uint64_t n);
+uint64_t reverse_bits(uint64_t n);
+uint64_t reverse_bits_limited(uint64_t n, uint64_t value);
 C_KZG_RET bit_reversal_permutation(void *values, size_t size, uint64_t n);
 void compute_powers(fr_t *out, const fr_t *x, uint64_t n);
 bool pairings_verify(const g1_t *a1, const g2_t *a2, const g1_t *b1, const g2_t *b2);
